Create Assignment5
The Assignment 5 class displays a menu of choices (add a product, compute the total cost, search a product, list product, quit, display menu) to a user. Then it performs the chosen task. It will keep asking a user to enter the next choice until the choice of 'Q' (Quit) is entered.